fixed some vc8 compiler warnings

+++ xreal/code/win32/win_main.c
@@ -591,12 +591,12 @@ extern char    *FS_BuildOSPath(const char *base, const char *game, const char *q
 // fqpath param added 7/20/02 by T.Ray - Sys_LoadDll is only called in vm.c at this time
 // fqpath will be empty if dll not loaded, otherwise will hold fully qualified path of dll module loaded
 // fqpath buffersize must be at least MAX_QPATH+1 bytes long
-void           *QDECL Sys_LoadDll(const char *name, char *fqpath, int (QDECL ** entryPoint) (int, ...),
-								  int (QDECL * systemcalls) (int, ...))
+void           *QDECL Sys_LoadDll(const char *name, char *fqpath, intptr_t (QDECL ** entryPoint) (int, ...),
+								  intptr_t (QDECL * systemcalls) (intptr_t, ...))
 {
 	static int      lastWarning = 0;
 	HINSTANCE       libHandle;
-	void            (QDECL * dllEntry) (int (QDECL * syscallptr) (int, ...));
+	void            (QDECL * dllEntry) (intptr_t (QDECL * syscallptr) (intptr_t, ...));
 	char           *basepath;
 	char           *cdpath;
 	char           *gamedir;
